Ignore Serial Version UID in emergency
A frequent problem when using daemon is making changes to the joblet config class while using the default serializer, causing serial version uid mismatches. We should change the default serializer globally to handle this more gracefully, but when this error occurs it brings all processing to a standstill. This change attempts to ignore the UID mismatch and deserailze anyway in case the changes happened to be serialization compatible. This prevents major backups in those cases.