Give up daemon lock aggressively
The daemon lock is only meant to lock on config production. Once
the config is obtained, it doesn't make sense to hold on to the
lock. Just let other daemons get the lock (and new configs!).

This issue manifests itself when aggressively scaling up daemons,
especially when running in kubernetes, where autoscaling is
common. In that case, when a large number of daemons are added,
all of them wait on the daemon lock while the daemon processes
the entire joblet. This doesn't really give us much, so just
be a good citizen and let go of the lock as soon as the config is
produced.